feat: implement live activities and refactor issues page
