+++ example/ios/TraceletWidget/AppIntent.swift
@@ -8,6 +8,7 @@
 import WidgetKit
 import AppIntents
 
+@available(iOS 17.0, *)
 struct ConfigurationAppIntent: WidgetConfigurationIntent {
     static var title: LocalizedStringResource { "Configuration" }
     static var description: IntentDescription { "This is an example widget." }
