+++ sdk/ios/Sources/TraceletSDK/ConfigManager.swift
@@ -148,8 +148,16 @@ public final class ConfigManager {
     public func getPreventSuspend() -> Bool { cache["preventSuspend"] as? Bool ?? false }
     public func getUseBackgroundActivitySession() -> Bool { cache["useBackgroundActivitySession"] as? Bool ?? false }
     
-    public func getLiveActivityConfig() -> TlLiveActivityConfig? {
-        return cache["liveActivityConfig"] as? TlLiveActivityConfig
+    public struct LiveActivityConfig {
+        public let title: String
+        public let body: String
+    }
+
+    public func getLiveActivityConfig() -> LiveActivityConfig? {
+        guard let map = cache["liveActivityConfig"] as? [String: Any],
+              let title = map["title"] as? String,
+              let body = map["body"] as? String else { return nil }
+        return LiveActivityConfig(title: title, body: body)
     }
 
     // MotionConfig
@@ -249,6 +257,8 @@ public final class ConfigManager {
     public func getEnableDeltaCompression() -> Bool { cache["enableDeltaCompression"] as? Bool ?? false }
     // Default must match the Dart HttpConfig default (5), not 6 (Issue #137).
     public func getDeltaCoordinatePrecision() -> Int { (cache["deltaCoordinatePrecision"] as? NSNumber)?.intValue ?? 5 }
+    public func getSyncTelematics() -> Bool { cache["syncTelematics"] as? Bool ?? false }
+    public func getTelematicsUrl() -> String { cache["telematicsUrl"] as? String ?? "" }
     
     public func getHttpParams() -> [String: Any] {
         if let params = cache["params"] as? [String: Any] {
